requirements.txt
telegram
configparser
redis
python-telegram-bot==13.7
psycopg2-binary
requests